# lib/openapi_3.1.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/go/api/admin/materials/svn/notify": {"post": {"tags": ["Materials"], "summary": "Materials Notify", "operationId": "materials_notify_go_api_admin_materials_svn_notify_post", "security": [{"HTTPBearer": []}], "parameters": [{"name": "repository_url", "in": "query", "required": true, "schema": {"type": "string", "title": "Repository Url"}}, {"name": "Accept", "in": "header", "required": false, "schema": {"type": "string", "description": "API Accept Header", "example": "application/vnd.go.cd.v2+json", "default": "application/vnd.go.cd.v2+json", "title": "Accept"}, "description": "API Accept Header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "string", "title": "Response Materials Notify Go Api Admin Materials Svn Notify Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/go/api/pipelines/{pipeline_name}/status": {"get": {"tags": ["Pipelines"], "summary": "Pipeline Status", "operationId": "pipeline_status_go_api_pipelines__pipeline_name__status_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "pipeline_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Pipeline Name"}}, {"name": "Accept", "in": "header", "required": false, "schema": {"type": "string", "description": "API Accept Header", "example": "application/vnd.go.cd.v1+json", "default": "application/vnd.go.cd.v1+json", "title": "Accept"}, "description": "API Accept Header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PipelineStatusResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/go/api/jobs/{pipeline_name}/{stage_name}/{job_name}/history": {"get": {"tags": ["Jobs"], "summary": "Pipeline Job History", "operationId": "pipeline_job_history_go_api_jobs__pipeline_name___stage_name___job_name__history_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "pipeline_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Pipeline Name"}}, {"name": "stage_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Stage Name"}}, {"name": "job_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Job Name"}}, {"name": "Accept", "in": "header", "required": false, "schema": {"type": "string", "description": "API Accept Header", "example": "application/vnd.go.cd.v1+json", "default": "application/vnd.go.cd.v1+json", "title": "Accept"}, "description": "API Accept Header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobHistoryResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "JobHistoryResponse": {"properties": {"jobs": {"items": {"$ref": "#/components/schemas/JobHistoryResponseItem"}, "type": "array", "title": "Jobs"}}, "type": "object", "required": ["jobs"], "title": "JobHistoryResponse"}, "JobHistoryResponseItem": {"properties": {"name": {"type": "string", "title": "Name"}, "state": {"$ref": "#/components/schemas/JobState"}, "result": {"type": "string", "title": "Result"}, "original_job_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Original Job Id"}, "scheduled_date": {"type": "string", "format": "date-time", "title": "Scheduled Date"}, "rerun": {"type": "boolean", "title": "Rerun"}, "agent_uuid": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Agent Uuid"}, "pipeline_name": {"type": "string", "title": "Pipeline Name"}, "pipeline_counter": {"type": "integer", "title": "Pipeline Counter"}, "stage_name": {"type": "string", "title": "Stage Name"}, "stage_counter": {"type": "string", "title": "Stage Counter"}}, "type": "object", "required": ["name", "state", "result", "original_job_id", "scheduled_date", "rerun", "agent_uuid", "pipeline_name", "pipeline_counter", "stage_name", "stage_counter"], "title": "JobHistoryResponseItem"}, "JobState": {"type": "string", "enum": ["Scheduled", "Assigned", "Preparing", "Building", "Completing", "Completed", "Rescheduled"], "title": "JobState"}, "PipelineStatusResponse": {"properties": {"paused": {"type": "boolean", "title": "Paused"}, "paused_cause": {"type": "string", "title": "Paused Cause"}, "paused_by": {"type": "string", "title": "Paused By"}, "locked": {"type": "boolean", "title": "Locked"}, "schedulable": {"type": "boolean", "title": "Schedulable"}}, "type": "object", "required": ["paused", "paused_cause", "paused_by", "locked", "schedulable"], "title": "PipelineStatusResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}